Remove less useful PopAnalysis Workflows
