+++ src/simmate/apps/warrenapp/workflows/__init__.py
@@ -8,19 +8,12 @@ from .nested_dft import (
     Relaxation__Warren__PbeWithWavecar,
 )
 from .population_analysis import (
-    PopulationAnalysis__Warren__BadelfHse,
-    PopulationAnalysis__Warren__BadelfPbesol,
+
     PopulationAnalysis__Warren__BaderBadelf,
     PopulationAnalysis__Warren__BaderBadelfHse,
     PopulationAnalysis__Warren__BaderBadelfPbesol,
-    PopulationAnalysis__Warren__BaderHse,
-    PopulationAnalysis__Warren__BaderPbesol,
-    PopulationAnalysis__Warren__ElfHse,
-    PopulationAnalysis__Warren__ElfPbesol,
     StaticEnergy__Warren__PrebadelfHse,
     StaticEnergy__Warren__PrebadelfPbesol,
-    StaticEnergy__Warren__PrebaderHse,
-    StaticEnergy__Warren__PrebaderPbesol,
 )
 from .relaxation import (
     Relaxation__Warren__Hse,
