+++ src/simmate/apps/warrenapp/workflows/population_analysis/__init__.py
@@ -1,17 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from .badelf_hse import PopulationAnalysis__Warren__BadelfHse
-from .badelf_pbesol import PopulationAnalysis__Warren__BadelfPbesol
-from .bader_badelf import PopulationAnalysis__Warren__BaderBadelf
-from .bader_badelf_hse import PopulationAnalysis__Warren__BaderBadelfHse
-from .bader_badelf_pbesol import PopulationAnalysis__Warren__BaderBadelfPbesol
-from .bader_hse import PopulationAnalysis__Warren__BaderHse
-from .bader_pbesol import PopulationAnalysis__Warren__BaderPbesol
-from .elf_hse import PopulationAnalysis__Warren__ElfHse
-from .elf_pbesol import PopulationAnalysis__Warren__ElfPbesol
-from .prebader_badelf_dft import (
+from .badelf import PopulationAnalysis__Warren__BaderBadelf
+from .badelf_hse import PopulationAnalysis__Warren__BaderBadelfHse
+from .badelf_pbesol import PopulationAnalysis__Warren__BaderBadelfPbesol
+from .prebadelf_dft import (
     StaticEnergy__Warren__PrebadelfHse,
     StaticEnergy__Warren__PrebadelfPbesol,
-    StaticEnergy__Warren__PrebaderHse,
-    StaticEnergy__Warren__PrebaderPbesol,
 )
