+++ src/simmate/apps/warrenapp/workflows/population_analysis/badelf_pbesol.py
@@ -1,17 +1,19 @@
 # -*- coding: utf-8 -*-
 
-from simmate.apps.warrenapp.workflows.population_analysis.base import VaspBadElfBase
-from simmate.apps.warrenapp.workflows.population_analysis.prebader_badelf_dft import (
+
+from simmate.engine.workflow import Workflow
+
+from simmate.apps.warrenapp.workflows.population_analysis.base import VaspBaderBadElfBase
+from simmate.apps.warrenapp.workflows.population_analysis.prebadelf_dft import (
     StaticEnergy__Warren__PrebadelfPbesol,
 )
 
 
-class PopulationAnalysis__Warren__BadelfPbesol(VaspBadElfBase):
+class PopulationAnalysis__Warren__BaderBadelfPbesol(VaspBaderBadElfBase):
     """
-    Runs a static energy calculation using an extra-fine FFT grid and then
-    carries out Bader analysis on the resulting charge density using the ELFCAR
-    as a reference when partitioning. Uses the PBE functional and settings from
-    the Warren Lab.
+    Runs a static energy calculation using an extra-fine FFT grid using vasp
+    and then carries out Badelf and Bader analysis on the resulting charge density.
+    Uses the Warren lab settings Pbesol settings.
     """
 
-    static_energy_prebadelf = StaticEnergy__Warren__PrebadelfPbesol
+    static_energy_prebadelf: Workflow = StaticEnergy__Warren__PrebadelfPbesol
