+++ src/simmate/apps/warrenapp/workflows/population_analysis/prebadelf_dft.py
@@ -3,14 +3,9 @@ from simmate.apps.warrenapp.workflows.static_energy.hse import StaticEnergy__War
 from simmate.apps.warrenapp.workflows.static_energy.pbesol import StaticEnergy__Warren__Pbesol
 
 # We want to define the settings that will be used when updating static energy
-# workflows for prebader or prebadelf DFT calculations. We do that here so
-# that we don't need to do it in every inheriting class.
-prebader_incar_settings = dict(
-    NGXF__density_a=10, # fine FFT grid (e.g. useds NGX instead of NGXF)
-    NGYF__density_b=10,
-    NGZF__density_c=10,
-    LAECHG=True,  # write core charge density to AECCAR0 and valence to AECCAR2
-)
+# workflows for prebadelf DFT calculations. We do that here so that we don't
+# need to do it in every inheriting class.
+
 prebadelf_incar_settings = dict(
     NGX__density_a=10,  # Note that these set the FFT grid while the pre-Bader task sets the
     NGY__density_b=10,  # fine FFT grid (e.g. useds NGX instead of NGXF)
@@ -20,39 +15,6 @@ prebadelf_incar_settings = dict(
     PREC="Single",  # ensures CHGCAR grid matches ELFCAR grid
     LAECHG=True,  # write core charge density to AECCAR0 and valence to AECCAR2
 )
-
-
-class StaticEnergy__Warren__PrebaderPbesol(StaticEnergy__Warren__Pbesol):
-    """
-    Runs a static energy calculation with a high-density FFT grid setting.
-    Results can be used for Bader analysis.
-
-    We do NOT recommend running this calculation on its own. Instead, you should
-    use the full workflow, which runs this calculation AND the following bader
-    analysis for you. This S3Task is only the first step of that workflow.
-    """
-
-    # The key thing for bader analysis is that we need a very fine FFT mesh. Other
-    # than that, it's the same as a static energy calculation.
-    incar = StaticEnergy__Warren__Pbesol.incar.copy()
-    incar.update(prebader_incar_settings)
-
-
-class StaticEnergy__Warren__PrebaderHse(StaticEnergy__Warren__Hse):
-    """
-    Runs a static energy calculation with a high-density FFT grid setting.
-    Results can be used for Bader analysis.
-
-    We do NOT recommend running this calculation on its own. Instead, you should
-    use the full workflow, which runs this calculation AND the following bader
-    analysis for you. This S3Task is only the first step of that workflow.
-    """
-
-    # The key thing for bader analysis is that we need a very fine FFT mesh. Other
-    # than that, it's the same as a static energy calculation.
-    incar = StaticEnergy__Warren__Hse.incar.copy()
-    incar.update(prebader_incar_settings)
-
 
 class StaticEnergy__Warren__PrebadelfPbesol(StaticEnergy__Warren__Pbesol):
     """
